pyproject.toml
[tool.poetry]
name = "testedapp"
version = "0.1.0"
description = ""
authors = ["Isfireomat <dfdfgfgf666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
